pyproject.toml
[tool.poetry]
name = "compatlib"
version = "0.1.0"
description = "utilities for compatibility with multiple versions"
authors = ["Tyler M Kontra"]
license = "MIT"
packages = [
    { include = "compatlib", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.6.2"
plum-dispatch = "^1.5.4"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
rope = "^0.20.1"
mypy = "^0.910"
black = "^21.9b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
